+++ src/components/component-detail/SessionDetail.js
